requirements.txt
discord==2.3.1
pyyaml==6.0